requirements.txt
# Database drivers
pymongo>=4.0.0
python-arango>=7.0.0
ravendb>=5.2.0

# Data processing
pandas>=1.5.0

# Environment variables
python-dotenv>=1.0.0

# Utilities
requests
psutil
matplotlib
